feat: Penências revisadas em 24/10/2025

+++ src/Main.java
@@ -15,15 +15,15 @@ public class Main {
         TabelaHash tabela1 = new TabelaHashMetodo1(CAPACIDADE_INICIAL, FATOR_CARGA);
         TabelaHash tabela2 = new TabelaHashMetodo2(CAPACIDADE_INICIAL, FATOR_CARGA);
 
-        long tIns1 = MedidorPerformance.medirTempoInsercao(tabela1, nomes);
-        long tIns2 = MedidorPerformance.medirTempoInsercao(tabela2, nomes);
+        double tIns1 = MedidorPerformance.medirTempoInsercao(tabela1, nomes);
+        double tIns2 = MedidorPerformance.medirTempoInsercao(tabela2, nomes);
 
         int qtdBusca = Math.min(100, total);
         String[] amostra = new String[qtdBusca];
         for (int i = 0; i < qtdBusca; i++) amostra[i] = nomes[i];
 
-        long tBus1 = MedidorPerformance.medirTempoBusca(tabela1, amostra);
-        long tBus2 = MedidorPerformance.medirTempoBusca(tabela2, amostra);
+        double tBus1 = MedidorPerformance.medirTempoBusca(tabela1, amostra);
+        double tBus2 = MedidorPerformance.medirTempoBusca(tabela2, amostra);
 
         RelatorioComparativo.gerar(
                 tabela1.getNomeFuncaoHash(), tabela1, tIns1, tBus1,
